Refactored the app to have the routes in separate files
The app.py was getting horribly messy. I've moved the API endpoints to a routes module, which can be imported into the main app. This doesn't have much computational overhead, and will make things more modular.
Later, we can add routes/omop.py and put database operations in there.
The ability to use prefixes on routes will mean we can rationalise API calls
